Add files via upload
My scratch project.
In my Scratch project, you control a sprite that moves in all four directions using the arrow keys. The goal is to eat only the fruits while avoiding cakes and donuts. The objects spawn on the right and move toward the left. If the fruits touch the left wall or if you eat a cake or donut, the game ends. Collect only the fruits to keep playing.